requirements.txt
# Core Streamlit & Web Framework
streamlit>=1.28.0
streamlit-authenticator>=0.2.3
streamlit-option-menu>=0.3.6

# Enterprise AI/NLP Stack
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
spacy>=3.7.0
nltk>=3.8.1
textblob>=0.17.1
openai>=1.3.0

# Advanced Document Processing
PyPDF2>=3.0.1
python-docx>=0.8.11
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Science & Visualization
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
altair>=5.1.0

# Professional PDF Generation
reportlab>=4.0.0
fpdf2>=2.7.0
weasyprint>=60.0

# Database & Authentication
sqlite3
bcrypt>=4.0.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Web APIs & Scraping
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Cloud & Deployment
boto3>=1.29.0
azure-storage-blob>=12.19.0
google-cloud-storage>=2.10.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
psutil>=5.9.0

# Development & Testing
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
sqlalchemy
Pillow
fastapi
uvicorn
python-jose
passlib

# NEW: Enhanced AI Features
sentence-transformers
textblob
scikit-learn
openai
reportlab

